[Python] Move more plugin logic to PythonPluginSystem
Part of #1445. In preparation for adding more plugin types, and since
we've committed to breaking the signature of
`PythonPluginSystem.scan()`, move more of the generic plugin loading
logic from `PythonPluginSystemManagerImplementationFactory` to
`PythonPluginSystem`.

Simplify tests to abuse the patchability of Python and be less reliant
on stub plugins on the file system.

Signed-off-by: David Feltell <david.feltell@foundry.com>