[UI] Add UI delegation API
Closes #1445.

Add a new `openassetio-ui` C++ library containing the core
implementation of the UI delegate system.

Include the `openassetio-ui` library in the existing Python bindings.
That is, there is no separate Python C module or distribution for the UI
delegation API. This greatly simplifies code reuse, and is a nod to the
"batteries included" philosophy of Python.

Add a `UIDelegateImplementationFactoryInterface` that mirrors the
`ManagerImplementationFactoryInterface`. That is, it is the abstract
base class providing the interface that hosts use to load UI delegate
plugins.

Signed-off-by: David Feltell <david.feltell@foundry.com>