squash: Add createPythonPluginSystemUIDelegateImplementationFactory
Add `createPythonPluginSystemUIDelegateImplementationFactory`, mirroring
`createPythonPluginSystemManagerImplementationFactory`.

This function gives access to the Python plugin system
(`PythonPluginSystemUIDelegateImplementationFactory`) from C++. The
Python plugin system is written entirely in pure Python, but inherits
from the bound `UIDelegateImplementationFactoryInterface` class. This
means we can query the Python interpreter for the Python plugin system,
and cast it to a C++ interface.

Signed-off-by: David Feltell <david.feltell@foundry.com>