squash: Add UIDelegate and UIDelegateFactory
Add `UIDelegate` middleware class, mirroring the `Manager` middleware
class. This class is the host-facing API that wraps the manager-provided
`UIDelegateInterface` implementation. Instances of this class are the
main point of contact between a host application and a UI delegate
plugin.

Add a `UIDelegateFactory` class, mirroring the `ManagerFactory`
class. The `UIDelegateFactory` is the host-facing API for searching,
identifying, configuring and constructing `UIDelegate` instances. In
particular, this class is responsible for parsing OpenAssetIO config
files and using their contents to initialise a UI delegate, retrieved
via a provided plugin system
(`UIDelegateImplementationFactoryInterface`).

The environment variable used to configure the location of a config file
is currently set the same for both manager plugins and UI delegate
plugins, meaning the same config file will be used to identify and
initialise both. Which manager and UI delegate plugin to instantiate is
determined by the (same) `identifier` value in the config file,
implicitly tying the two plugins together. However, the settings dict
for the manager and UI delegate are separated into separate TOML
sections.

Signed-off-by: David Feltell <david.feltell@foundry.com>