squash: Add UIDelegateRequestInterface and UIDelegateStateInterface
Add `UIDelegateRequest` for encapsulating the part of a request for UI
delegation that may change over time. For example, this could include
the selection of entities that the host wishes the UI delegate to
present additional information for. Such a list could change as the user
interacts with the other parts of the host whilst the delegated UI
element is visible.

Similarly, add a `UIDelegateState` for encapsulating the current state
of a UI delegate. For example, the currently selected entities in a
delegated browser.

As is common across the API, split these types into interface and
middleware classes, namespaced according to whose responsibility it is
for providing vs. using. I.e. `UIDelegateStateInterface` is implemented
and constructed by the UI delegate, and so is under `managerApi`;
whereas `UIDelegateState` is the host-facing middleware, so is under
`hostApi`. Similarly, `UIDelegateRequestInterface` is under `hostApi`
and `UIDelegateRequest` is under `managerApi`.

Using this separation of middleware and interface allows functionality
to be injected in-between, e.g. validation, and provides some level of
isolation of one side of the API in case of changes or additions to the
other side.

The request and state objects are also the mechanism for establishing
communication channels with an ongoing UI delegation request. This is
via the `stateChangedCallback` on the request, which is called by the UI
delegate when its (host-relevant) state has changed; and the
`updateRequestCallback` on the state, which is called by the host when
it wishes to update the parameters of the original request.

To remain agnostic to any particular UI framework, both the request and
state hold an arbitrary `nativeData` object as a `std::any`. This is
typically where a widget object would be stored. But there are other
possibilities, for example hosts that use an immediate mode GUI
framework could use `nativeData` to store a function that is called
every frame. It is up to the host application to document what is/should
be placed in `nativeData` and under what conditions. Documentation in
UI-specific OpenAssetIO trait libraries can provide further guidance on
this.

When a C++ host or UI delegate implementation wishes to interact with a
Python UI delegate or host, respectively, then the Python bindings for
the `nativeData()` functions assume a `PyObject*` is stored in the
`std::any`. It is up to the C++ side to further marshal data in/out of
the PyObject. For example, in Qt the Shiboken API can be used to convert
QWidget objects to/from PyObject objects.

Signed-off-by: David Feltell <david.feltell@foundry.com>