squash: Add UIAccess
Add new enum class, `UIAccess`, similar to other enum classes (e.g.
`ResolveAccess`, `PublishingAccess`, etc) to allow UI delegation
requests to be further parametrised by the host's desired access
pattern. For example, given that the host wants a browser of some kind
and provides an entity reference, the access mode can be used to
disambiguate whether that browser is for searching for related entities,
or for publishing to the entity, in an analogy to "File->Open" vs.
"File->Save As" menus in traditional file-based workflows.

Signed-off-by: David Feltell <david.feltell@foundry.com>