squash: Add populateUI
Add a `populateUI` method to the `UIDelegate` (and corresponding
`UIDelegateInterface`) classes. This is the core method used to delegate
UI elements. It is heavily parametrised by UI-specific traits, access
mode, and a `UIDelegateRequest`, as well as the usual `Context` and
`HostSession`.

Combined, these parameters provide all the information required to
determine the nature of the UI to create, how it should be populated in
the host, and how changes can be communicated asynchronously.

It is expected that UI-specific traits will detail

- The kind of UI to create (e.g. browser vs. inline control vs. menu).
- Some details on where to create it (e.g. a new widget placed in the
  returned state's `nativeData` vs. mutations to a widget in the initial
  request's `nativeData` vs. a temporary modal dialog).
- The data required by the host (e.g. single vs. multiple items; entity
  references vs. entity traits vs. both).
- The expected lifetime of the UI (e.g. single-use vs. persistent).

The UI traits will be defined in a separate industry-specific repository
(e.g. OpenAssetIO-MediaCreation), and their documentation must be
consulted to interpret their meaning.

The host's own documentation must be queried for more specifics beyond
what these host/framework-agnostic traits can provide.

The provided access mode helps to further bifurcate the type of UI
presented, by specifying whether the UI is for reading or publishing.
This could radically alter the nature of the UI element, even for the
same set of UI traits.

The `UIDelegateRequest` and `UIDelegateState` (and their corresponding
`Interface`s) are used to establish asynchronous communication during
an ongoing UI delegation request. They both have an (optional) callback
hook, allowing the initial state or request to be updated, respectively.

Signed-off-by: David Feltell <david.feltell@foundry.com>