squash: Add close()
Add a `close()` method that allows the UI delegate plugin to clean up
when all UI elements associated with it are destroyed wholesale. For
example, when quitting the host application, or when the host closes a
document to open a new one.

Call the wrapped `UIDelegateInterface` implementation's `close()` method
on destruction of the `UIDelegate` middleware. This provides plugins
with a reliable clean-up hook, especially in the case of Python plugins,
where overriding `__del__` is considered an anti-pattern since it can
prevent garbage collection.

This mirrors a similar mechanism in the `EntityReferencePager` class.

Note that, in general, pybind11 calls C++ destructors with the GIL held,
whereas we always default to releasing the GIL when calling C++
functions, so we have a bit of an inconsistency. Recent versions of
pybind11 have a `release_gil_before_calling_cpp_dtor` tag that can be
added to binding definitions. However, we must be careful about the
minimum supported version of pybind11, since it is not a truly private
dependency - see #1420. So rather than bump the pybind11 version, just
document a warning that the GIL will be held when `close()` is called
on destruction, for now.

Signed-off-by: David Feltell <david.feltell@foundry.com>